refactor(deploy): split deploy.rs into staging/injection/rewrite/filtering submodules
Mirrors the add.rs split (#740) shape: 2118-line file → mod.rs + four
sibling modules with tests colocated.

  - mod.rs (485)        DeployArgs, run, seed_target_lockfile, canonicalize
  - staging.rs (256)    stage_one, ensure_target_writable, collect_all_files
  - injection.rs (315)  Injection, plan_injections, materialize_injections
  - rewrite.rs (934)    rewrite_local_refs, catalog/workspace spec resolution
  - filtering.rs (245)  StripFields, DepAxis, dep_selection_for_args

External surface (`commands::deploy::DeployArgs`, `commands::deploy::run`)
unchanged. All 29 unit tests pass.
